Rename FileButtonTextScss
Rename it to its correct name, update imports and references
